created user model, controller protected goals api routes

+++ backend/controllers/goalController.js
@@ -1,18 +1,19 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 //  @desc Get goals
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find();
+    const goals = await Goal.find({ user: req.user.id });
     res.status(200).json(goals);
 });
 
 //  @desc Set goal
 // @route POST /api/goals
 // @access Private
-const setGoals = asyncHandler(async (req, res) => {
+const setGoal = asyncHandler(async (req, res) => {
     if (!req.body.text) {
         res.status(400);
         throw new Error("please add a text field");
@@ -20,6 +21,7 @@ const setGoals = asyncHandler(async (req, res) => {
 
     const goal = await Goal.create({
         text: req.body.text,
+        user: req.user.id,
     });
     res.status(200).json(goal);
 });
@@ -29,10 +31,20 @@ const setGoals = asyncHandler(async (req, res) => {
 // @access Private
 const updateGoal = asyncHandler(async (req, res) => {
     const goal = await Goal.findById(req.params.id);
-
     if (!goal) {
         res.status(400);
         throw new Error("Goal not found");
+    }
+
+    const user = await User.findById(req.user.id);
+    if (!user) {
+        res.status(401);
+        throw new Error("User not found");
+    }
+
+    if (goal.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error("User not authorized");
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -50,6 +62,17 @@ const deleteGoal = asyncHandler(async (req, res) => {
         res.status(400);
         throw new Error("Goal not found");
     }
+
+    const user = await User.findById(req.user.id);
+    if (!user) {
+        res.status(401);
+        throw new Error("User not found");
+    }
+
+    if (goal.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error("User not authorized");
+    }
     const deletedGoal = await Goal.findByIdAndDelete(req.params.id);
     res.status(200).json({ message: `Deleted goal ${deletedGoal}` });
 });
@@ -57,6 +80,6 @@ const deleteGoal = asyncHandler(async (req, res) => {
 module.exports = {
     getGoals,
     deleteGoal,
-    setGoals,
+    setGoal,
     updateGoal,
 };
